+++ lib/data/constants/k_values.dart
@@ -1,0 +1,2 @@
+const kValuesTimeoutDuration = Duration.zero;
+const kValuesMinBusyDuration = Duration.zero;